# [refactor] Refactoring of hyperscript.js and render.js, including performance improvements