--- a/util/assign.js
+++ b/util/assign.js
@@ -1,4 +1,0 @@
-// This exists so I'm only saving it once.
-"use strict"
-
-module.exports = Object.assign