<!--- Provide a general summary of your changes in the Title above -->
Refactor hyperscript.js and render.js. In particular, the replacement of fix #2622 appears to have significantly improved the performance regression.

## Description
<!--- Describe your changes in detail -->
This change is a refactoring aimed at improving code outlook and performance.
- hyperscript.js now uses Object.assign() to merge attrs. Also, Object.assign() controls the order in which attributes are applied, which addresses issue #2622 alternatively.
- In render.js, performance was improved by eliminating the past #2622 fix. Also, isFileInput has been moved into setAttr() to improve code outlook.
- Polyfill for Object.assign() is removed. The polyfill doesn't support multiple source objects (like in this pr), and almost all browsers now natively support Object.assign().

<details>
<summary>results of `npm run perf` </summary>

before (v2.2.8)
> construct large vnode tree x 18,987 ops/sec ±0.79% (126 runs sampled)
> rerender identical vnode x 3,898,180 ops/sec ±0.52% (127 runs sampled)
> rerender same tree x 55,444 ops/sec ±1.71% (128 runs sampled)
> add large nested tree x 6,701 ops/sec ±1.22% (123 runs sampled)
> mutate styles/properties x 78.77 ops/sec ±3.08% (75 runs sampled)
> repeated add/removal x 3,052 ops/sec ±1.15% (126 runs sampled)
> Completed perf tests in 60100ms

after ("rerender same tree" is improved)
> construct large vnode tree x 19,972 ops/sec ±0.81% (123 runs sampled)
> rerender identical vnode x 3,889,751 ops/sec ±1.03% (127 runs sampled)
> rerender same tree x 88,338 ops/sec ±0.27% (125 runs sampled)
> add large nested tree x 7,014 ops/sec ±1.61% (119 runs sampled)
> mutate styles/properties x 77.55 ops/sec ±1.12% (73 runs sampled)
> repeated add/removal x 3,067 ops/sec ±2.48% (90 runs sampled)
> Completed perf tests in 59057ms

cf. v2.0.4 (by backported test-perf.js and m.censor)
> construct large vnode tree x 13,149 ops/sec ±0.51% (124 runs sampled)
> rerender identical vnode x 5,084,840 ops/sec ±0.49% (127 runs sampled)
> rerender same tree x 82,781 ops/sec ±0.18% (128 runs sampled)
> add large nested tree x 6,611 ops/sec ±2.56% (101 runs sampled)
> mutate styles/properties x 79.76 ops/sec ±0.86% (73 runs sampled)
> repeated add/removal x 3,067 ops/sec ±0.97% (125 runs sampled)
> Completed perf tests in 59579ms

</details>

As shown above, `npm run perf` is improved, especially for "rerender same tree".

I have also run test-perf.js in a real browser to see the performance improvement.
(If necessary, I would like to measure and tabulate performance later.)

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
As for performance, it is sufficiently sophisticated at v2.0.4. However, subsequent bug fixes and enhancements seem to have caused some performance regression. I measured the performance of v2.0.4 and later versions with `npm run perf` and found that https://github.com/MithrilJS/mithril.js/pull/2578/commits/7c7d76d524e6b4c4df2eec5a6f22935f7a400d8f of #2578 has a negative impact on performance.
The commit https://github.com/MithrilJS/mithril.js/pull/2578/commits/7c7d76d524e6b4c4df2eec5a6f22935f7a400d8f fixes #2622 by changing setAttr(s), but I noticed that it can be resolved by hyperscript(execSelector) instead of setAttr(s).
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
- `npm run test`
- I also confirmed with flems that #2622 is not reproduced.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [ ] My change requires a documentation update, and I've opened a pull request to update it already:
- [x] I have read https://mithril.js.org/contributing.html.