@kfule 2024-10-30
@dead-claudia
Thanks for your review.
I removed the assign.js module as you suggested.
Also, I removed unnecessary parameters from setAttr().